# src/docs/openapi.yaml
openapi: 3.0.0
info:
    title: WaitLess API
    version: 1.0.0
    description: |
        WaitLess is a Queue Management System REST API that lets users to join virtual queues, track their position in real time, receive notifications and lets staff/admin manage service centers and queues.


        ## Features
        - JWT Cookie Authentication
        - Role-Based Access Control (Admin, Staff & User)
        - Queue Management (join, check-in, call next, priority, stats)
        - Service Center Management
        - Staff assignment to service centers
        - Notification System
        - Queue Analytics
        - Nepal time formatted timestamps

        ## Technology Stack
        - Node.js
        - Express.js
        - PostgreSQL
        - Prisma ORM

        ## Authentication
        WaitLess uses a JWT stored inside an **HttpOnly Cookie** for authentication.

        After a successful login, the server automatically sets a secure `jwt` cookie. All subsequent authenticated requests automatically include this cookie, so there is no need to manually send an Authorization header.

        ## Response Format
        Every response (success or error) follows the same envelope: `success`, `statusCode`, `message` and `data` (or `errors` on failure).

    contact:
        name: Rupesh Choudhary
        email: rupevilary1010@gmail.com
servers:
    - url: http://localhost:8000/api/v1
      description: Local Development Server
...
      description: Production Server
...
security:
    - CookieAuth: []
...
paths:
...
    /service-centers:
        get:
            tags: [Service Center]
            summary: Get all service centers
            description: "Returns every service center, grouped by owner. Role allowed: ADMIN."
            responses:
                "200":
                    description: Service centers fetched successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/ApiResponse"
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: "#/components/schemas/ServiceCenterOwnerGroup"
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "403":
                    $ref: "#/components/responses/Forbidden"
...
    /staff-assignment:
        get:
            tags: [Staff Assignment]
            summary: Get all staff assignments (flat list)
            description: "Returns every staff assignment as a flat list. Role allowed: ADMIN."
            responses:
                "200":
                    description: Staff assignment list fetched successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/ApiResponse"
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: "#/components/schemas/StaffAssignment"
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "403":
                    $ref: "#/components/responses/Forbidden"
...
    /notification:
        get:
            tags: [Notification]
            summary: Get my notifications
            responses:
                "200":
                    description: Notifications fetched successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/ApiResponse"
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  unreadCount:
                                                      type: integer
                                                  totalNotifications:
                                                      type: integer
                                                  notifications:
                                                      type: array
                                                      items:
                                                          $ref: "#/components/schemas/Notification"
                "401":
                    $ref: "#/components/responses/Unauthorized"

        delete:
            tags: [Notification]
            summary: Delete all notifications
            responses:
                "200":
                    description: All notifications deleted successfully.
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "404":
                    description: No notifications found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApiError"
...
components:
    securitySchemes:
        CookieAuth:
            type: apiKey
            in: cookie
            name: jwt
            description: HttpOnly JWT cookie set automatically after login/registration
...
    responses:
...
        Unauthorized:
            description: Authentication required or the JWT cookie is missing/invalid.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ApiError"
        Forbidden:
            description: You do not have permission to perform this action.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ApiError"
...
    schemas:
        ApiResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                statusCode:
                    type: integer
                    example: 200
                message:
                    type: string
                    example: Request completed successfully.
                data:
                    description: Endpoint-specific response data.
                    nullable: true
        ApiError:
            type: object
            properties:
                success:
                    type: boolean
                    example: false
                statusCode:
                    type: integer
                    example: 400
                message:
                    type: string
                    example: Validation failed.
                errors:
                    type: array
                    items:
                        type: string
                    example:
                        - Email is required
                        - Password must be at least 8 characters

        # ---- User ----
...
        ServiceCenter:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                email:
                    type: string
                    format: email
                phoneNumber:
                    type: string
                address:
                    type: string
                logo:
                    type: string
                    nullable: true
                openingHour:
                    type: integer
                openingMinute:
                    type: integer
                closingHour:
                    type: integer
                closingMinute:
                    type: integer
                averageServiceTime:
                    type: integer
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                createdAtNepalTime:
                    type: string
                    description: createdAt formatted in Nepal time.
                updatedAtNepalTime:
                    type: string
                    description: updatedAt formatted in Nepal time.
...
        ServiceCenterOwnerGroup:
            type: object
            description: A service center owner along with the centers they own.
            properties:
                id:
                    type: string
                    format: uuid
                fullName:
                    type: string
                email:
                    type: string
                role:
                    type: string
                    enum: [USER, STAFF, ADMIN]
                serviceCenters:
                    type: array
                    items:
                        $ref: "#/components/schemas/ServiceCenter"

        # ---- Staff Assignment ----
...
        StaffAssignment:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                userId:
                    type: string
                    format: uuid
                serviceCenterId:
                    type: string
                    format: uuid
                user:
                    type: object
                    properties:
                        id:
                            type: string
                            format: uuid
                        fullName:
                            type: string
                        email:
                            type: string
                serviceCenter:
                    type: object
                    properties:
                        id:
                            type: string
                            format: uuid
                        name:
                            type: string
...
        Notification:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                userId:
                    type: string
                    format: uuid
                title:
                    type: string
                    example: Queue joined
                message:
                    type: string
                    example: You joined Nirjal Service Center. Your token number is 12.
                isRead:
                    type: boolean
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time